Moved the circuits to better reflect which board each circuit will be on
The only schematics edited were the sensor, microcontroller, and power schematics. I have saved the sensor schematic sheet locally to my computer, not that I think we would need it, but it is there. I checked all the circuits after I copied and pasted them and I believe they are all the same as before and I also double checked that none of the footprints were deleted or removed.